Fixed xterm/screen titles once and for all.
Both always show the full command name, or 'zsh $PWD'.

If I'm in a screen session OR on an SSH connection, xterm title includes
the current hostname.  If I'm not eevee/amunroe, the username is also
included.

screen windows do not include hostname, but do show the username if the
shell isn't owned by me.